+++ outputs.tf
@@ -3,7 +3,7 @@ output "namespace" {
   value       = helm_release.helm_release.metadata.0.namespace
 }
 
-output "app_version" {
+output "version" {
   description = "Application version"
   value       = helm_release.helm_release.metadata.0.app_version
 }