+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added recommanded VSCode extentions
 - Added some VSCode tasks
 
+### Changed
+
+- Rename `app_version` output variable to `version`
+
 ## [0.4.1] 2022-12-21
 
 ### Fixed